golang_str_pickle: Test it wrt all pickle modules we care about
We want bstr/ustr pickling support to be robust. So we need to test it
against all pickle modules that are in use. This includes python pickle
version from stdlib (pickle.py), C pickle version from stdlib (cPickle
on py2 and _pickle on py3) and, correspondingly, py and C versions from
zodbpickle.

-> Adjust pickling tests to cover all those variants.